Added top 50 sites from bro if http_agents not active
Added an else condition on the check for http_agents to display to 50 URLs.

If number of http_agents is not greater than 0 else condition will display top 50 sites from the previous day from Bro logs. Not sure this is the best way to get yesterday's date.

Also I wasn't sure what the best way to check that Bro is active is, that should probably be added in if this is accepted.