Remove redundant source call
At the top of the script is a call to
'source /etc/nsm/securityonion.conf' however, later in the script
when there's a check if ELSA is defined, another call to source
/etc/nsm/securityonion.conf is done (along with check to see if the
file exists first).  I simply moved the file check code up to the top
of the script and simplified the if statement that checks if ELSA
is defined.