Enable nullglobs to prevent string literal bug in various for loops
Take this code as an example:

for i in /nsm/sensor_data/*/stats.log; do
        echo "$i"
        if [ $( tail -n 50 $i | grep -c drop ) -ne 0 ]; then

If there are no files that match '/nsm/sensor_data/*/status.log', the
non-intuitive default setting in shell scripts is to assign the literal
'/nsm/sensor_data/*/status.log' to $i.  This then breaks the following if
statement because tail is being passed a filename with a wildcard in it.